Get single price ticker
Changed Symbol price ticker endpoint from v1 to v1. If symobl argument is not exist, all symbols will be returned in array. If symbole argument is exist, only return one value.
You must adjust the difference between the current time on the client and the time on the server.

See also
  - https://github.com/binance-exchange/binance-official-api-docs/blob/master/rest-api.md